Dashboards: Refactor dynamic dashboards code
