+++ public/app/features/dashboard-scene/scene/layout-responsive-grid/ResponsiveGridLayoutManager.tsx
@@ -1,6 +1,4 @@
-import { SelectableValue } from '@grafana/data';
 import { SceneComponentProps, SceneCSSGridLayout, SceneObjectBase, SceneObjectState, VizPanel } from '@grafana/scenes';
-import { Select } from '@grafana/ui';
 import { t } from 'app/core/internationalization';
 import { OptionsPaneItemDescriptor } from 'app/features/dashboard/components/PanelEditor/OptionsPaneItemDescriptor';
 
@@ -10,6 +8,7 @@ import { RowsLayoutManager } from '../layout-rows/RowsLayoutManager';
 import { DashboardLayoutManager } from '../types/DashboardLayoutManager';
 
 import { ResponsiveGridItem } from './ResponsiveGridItem';
+import { getEditOptions } from './ResponsiveGridLayoutManagerEditor';
 
 interface ResponsiveGridLayoutManagerState extends SceneObjectState {
   layout: SceneCSSGridLayout;
@@ -19,6 +18,8 @@ export class ResponsiveGridLayoutManager
   extends SceneObjectBase<ResponsiveGridLayoutManagerState>
   implements DashboardLayoutManager
 {
+  public static Component = ResponsiveGridLayoutManagerRenderer;
+
   public readonly isDashboardLayoutManager = true;
 
   public static readonly descriptor = {
@@ -37,13 +38,11 @@ export class ResponsiveGridLayoutManager
   public constructor(state: ResponsiveGridLayoutManagerState) {
     super(state);
 
-    //@ts-ignore
+    // @ts-ignore
     this.state.layout.getDragClassCancel = () => 'drag-cancel';
   }
 
-  public editModeChanged(isEditing: boolean): void {}
-
-  public addPanel(vizPanel: VizPanel): void {
+  public addPanel(vizPanel: VizPanel) {
     const panelId = dashboardSceneGraph.getNextPanelId(this);
 
     vizPanel.setState({ key: getVizPanelKeyForPanelId(panelId) });
@@ -54,18 +53,12 @@ export class ResponsiveGridLayoutManager
     });
   }
 
-  public addNewRow(): void {
-    const rowsLayout = RowsLayoutManager.createFromLayout(this);
-    rowsLayout.addNewRow();
-    getDashboardSceneFor(this).switchLayout(rowsLayout);
-  }
-
   public removePanel(panel: VizPanel) {
     const element = panel.parent;
     this.state.layout.setState({ children: this.state.layout.state.children.filter((child) => child !== element) });
   }
 
-  public duplicatePanel(panel: VizPanel): void {
+  public duplicatePanel(panel: VizPanel) {
     const gridItem = panel.parent;
     if (!(gridItem instanceof ResponsiveGridItem)) {
       console.error('Trying to duplicate a panel that is not inside a DashboardGridItem');
@@ -103,11 +96,25 @@ export class ResponsiveGridLayoutManager
     return panels;
   }
 
-  public getOptions(): OptionsPaneItemDescriptor[] {
-    return getOptions(this);
+  public addNewRow() {
+    const rowsLayout = RowsLayoutManager.createFromLayout(this);
+    rowsLayout.addNewRow();
+    getDashboardSceneFor(this).switchLayout(rowsLayout);
   }
 
-  public static createEmpty() {
+  public getOptions(): OptionsPaneItemDescriptor[] {
+    return getEditOptions(this);
+  }
+
+  public changeColumns(columns: string) {
+    this.state.layout.setState({ templateColumns: columns });
+  }
+
+  public changeRows(rows: string) {
+    this.state.layout.setState({ autoRows: rows });
+  }
+
+  public static createEmpty(): ResponsiveGridLayoutManager {
     return new ResponsiveGridLayoutManager({
       layout: new SceneCSSGridLayout({
         children: [],
@@ -125,96 +132,13 @@ export class ResponsiveGridLayoutManager
       children.push(new ResponsiveGridItem({ body: panel.clone() }));
     }
 
-    return new ResponsiveGridLayoutManager({
-      layout: new SceneCSSGridLayout({
-        children,
-        templateColumns: 'repeat(auto-fit, minmax(400px, auto))',
-        autoRows: 'minmax(300px, auto)',
-      }),
-    });
+    const layoutManager = ResponsiveGridLayoutManager.createEmpty();
+    layoutManager.state.layout.setState({ children });
+
+    return layoutManager;
   }
-
-  /**
-   * Might as well implement this as a no-top function as vs-code very eagerly adds optional functions that throw Method not implemented
-   */
-  public activateRepeaters(): void {}
-
-  public static Component = ({ model }: SceneComponentProps<ResponsiveGridLayoutManager>) => {
-    return <model.state.layout.Component model={model.state.layout} />;
-  };
 }
 
-function getOptions(layoutManager: ResponsiveGridLayoutManager): OptionsPaneItemDescriptor[] {
-  const options: OptionsPaneItemDescriptor[] = [];
-
-  const cssLayout = layoutManager.state.layout;
-
-  const rowOptions: Array<SelectableValue<string>> = [];
-  const sizes = [100, 150, 200, 250, 300, 350, 400, 450, 500, 550, 650];
-  const colOptions: Array<SelectableValue<string>> = [
-    { label: t('dashboard.responsive-layout.options.one-column', '1 column'), value: `1fr` },
-    { label: t('dashboard.responsive-layout.options.two-columns', '2 columns'), value: `1fr 1fr` },
-    { label: t('dashboard.responsive-layout.options.three-columns', '3 columns'), value: `1fr 1fr 1fr` },
-  ];
-
-  for (const size of sizes) {
-    colOptions.push({
-      label: t('dashboard.responsive-layout.options.min', 'Min: {{size}}px', { size }),
-      value: `repeat(auto-fit, minmax(${size}px, auto))`,
-    });
-  }
-
-  for (const size of sizes) {
-    rowOptions.push({
-      label: t('dashboard.responsive-layout.options.min', 'Min: {{size}}px', { size }),
-      value: `minmax(${size}px, auto)`,
-    });
-  }
-
-  for (const size of sizes) {
-    rowOptions.push({
-      label: t('dashboard.responsive-layout.options.fixed', 'Fixed: {{size}}px', { size }),
-      value: `${size}px`,
-    });
-  }
-
-  options.push(
-    new OptionsPaneItemDescriptor({
-      title: t('dashboard.responsive-layout.options.columns', 'Columns'),
-      render: () => {
-        const { templateColumns } = cssLayout.useState();
-        return (
-          <Select
-            options={colOptions}
-            value={String(templateColumns)}
-            onChange={(value) => {
-              cssLayout.setState({ templateColumns: value.value });
-            }}
-            allowCustomValue={true}
-          />
-        );
-      },
-    })
-  );
-
-  options.push(
-    new OptionsPaneItemDescriptor({
-      title: t('dashboard.responsive-layout.options.rows', 'Rows'),
-      render: () => {
-        const { autoRows } = cssLayout.useState();
-        return (
-          <Select
-            options={rowOptions}
-            value={String(autoRows)}
-            onChange={(value) => {
-              cssLayout.setState({ autoRows: value.value });
-            }}
-            allowCustomValue={true}
-          />
-        );
-      },
-    })
-  );
-
-  return options;
+function ResponsiveGridLayoutManagerRenderer({ model }: SceneComponentProps<ResponsiveGridLayoutManager>) {
+  return <model.state.layout.Component model={model.state.layout} />;
 }
