+++ public/app/features/dashboard-scene/scene/layout-rows/RowsLayoutManager.tsx
@@ -1,34 +1,24 @@
-import { css } from '@emotion/css';
-
-import { GrafanaTheme2 } from '@grafana/data';
-import {
-  SceneComponentProps,
-  sceneGraph,
-  SceneGridItemLike,
-  SceneGridRow,
-  SceneObjectBase,
-  SceneObjectState,
-  VizPanel,
-} from '@grafana/scenes';
-import { useStyles2 } from '@grafana/ui';
+import { SceneGridItemLike, SceneGridRow, SceneObjectBase, SceneObjectState, VizPanel } from '@grafana/scenes';
 import { t } from 'app/core/internationalization';
 
 import { isClonedKey } from '../../utils/clone';
-import { DashboardScene } from '../DashboardScene';
+import { dashboardSceneGraph } from '../../utils/dashboardSceneGraph';
 import { DashboardGridItem } from '../layout-default/DashboardGridItem';
 import { DefaultGridLayoutManager } from '../layout-default/DefaultGridLayoutManager';
 import { RowRepeaterBehavior } from '../layout-default/RowRepeaterBehavior';
-import { ResponsiveGridLayoutManager } from '../layout-responsive-grid/ResponsiveGridLayoutManager';
 import { DashboardLayoutManager } from '../types/DashboardLayoutManager';
 
 import { RowItem } from './RowItem';
 import { RowItemRepeaterBehavior } from './RowItemRepeaterBehavior';
+import { RowLayoutManagerRenderer } from './RowsLayoutManagerRenderer';
 
 interface RowsLayoutManagerState extends SceneObjectState {
   rows: RowItem[];
 }
 
 export class RowsLayoutManager extends SceneObjectBase<RowsLayoutManagerState> implements DashboardLayoutManager {
+  public static Component = RowLayoutManagerRenderer;
+
   public readonly isDashboardLayoutManager = true;
 
   public static readonly descriptor = {
@@ -44,47 +34,21 @@ export class RowsLayoutManager extends SceneObjectBase<RowsLayoutManagerState> i
 
   public readonly descriptor = RowsLayoutManager.descriptor;
 
-  public editModeChanged(isEditing: boolean): void {}
-
-  public addPanel(vizPanel: VizPanel): void {
+  public addPanel(vizPanel: VizPanel) {
     // Try to add new panels to the selected row
-    const selectedObject = this.getSelectedObject();
-    if (selectedObject instanceof RowItem) {
-      return selectedObject.onAddPanel(vizPanel);
+    const selectedRows = dashboardSceneGraph.getAllSelectedObjects(this).filter((obj) => obj instanceof RowItem);
+    if (selectedRows.length > 0) {
+      return selectedRows.forEach((row) => row.getLayout().addPanel(vizPanel));
     }
 
     // If we don't have selected row add it to the first row
     if (this.state.rows.length > 0) {
-      return this.state.rows[0].onAddPanel(vizPanel);
+      return this.state.rows[0].getLayout().addPanel(vizPanel);
     }
 
     // Otherwise fallback to adding a new row and a panel
     this.addNewRow();
-    this.state.rows[this.state.rows.length - 1].onAddPanel(vizPanel);
-  }
-
-  public addNewRow(): void {
-    this.setState({
-      rows: [
-        ...this.state.rows,
-        new RowItem({
-          title: t('dashboard.rows-layout.row.new', 'New row'),
-          layout: ResponsiveGridLayoutManager.createEmpty(),
-        }),
-      ],
-    });
-  }
-
-  public removePanel(panel: VizPanel) {}
-
-  public removeRow(row: RowItem) {
-    this.setState({
-      rows: this.state.rows.filter((r) => r !== row),
-    });
-  }
-
-  public duplicatePanel(panel: VizPanel): void {
-    throw new Error('Method not implemented.');
+    this.state.rows[this.state.rows.length - 1].getLayout().addPanel(vizPanel);
   }
 
   public getVizPanels(): VizPanel[] {
@@ -98,33 +62,42 @@ export class RowsLayoutManager extends SceneObjectBase<RowsLayoutManagerState> i
     return panels;
   }
 
-  public getOptions() {
-    return [];
+  public addNewRow() {
+    this.setState({ rows: [...this.state.rows, new RowItem()] });
+  }
+
+  public editModeChanged(isEditing: boolean) {
+    this.state.rows.forEach((row) => row.getLayout().editModeChanged?.(isEditing));
   }
 
   public activateRepeaters() {
     this.state.rows.forEach((row) => {
-      if (row.state.$behaviors) {
-        for (const behavior of row.state.$behaviors) {
-          if (behavior instanceof RowItemRepeaterBehavior && !row.isActive) {
-            row.activate();
-            break;
-          }
+      const behavior = (row.state.$behaviors ?? []).find((b) => b instanceof RowItemRepeaterBehavior);
+
+      if (behavior) {
+        if (!row.isActive) {
+          row.activate();
         }
 
         if (!row.getLayout().isActive) {
           row.getLayout().activate();
         }
+
+        if (!behavior.isActive) {
+          behavior.activate();
+        }
       }
     });
   }
 
-  public getSelectedObject() {
-    return sceneGraph.getAncestor(this, DashboardScene).state.editPane.state.selection?.getFirstObject();
+  public removeRow(row: RowItem) {
+    this.setState({
+      rows: this.state.rows.filter((r) => r !== row),
+    });
   }
 
-  public static createEmpty() {
-    return new RowsLayoutManager({ rows: [] });
+  public static createEmpty(): RowsLayoutManager {
+    return new RowsLayoutManager({ rows: [new RowItem()] });
   }
 
   public static createFromLayout(layout: DashboardLayoutManager): RowsLayoutManager {
@@ -175,7 +148,7 @@ export class RowsLayoutManager extends SceneObjectBase<RowsLayoutManagerState> i
       rows = config.map(
         (rowConfig) =>
           new RowItem({
-            title: rowConfig.title ?? t('dashboard.rows-layout.row.new', 'New row'),
+            title: rowConfig.title,
             isCollapsed: !!rowConfig.isCollapsed,
             layout: DefaultGridLayoutManager.fromGridItems(
               rowConfig.children,
@@ -186,34 +159,9 @@ export class RowsLayoutManager extends SceneObjectBase<RowsLayoutManagerState> i
           })
       );
     } else {
-      rows = [new RowItem({ layout: layout.clone(), title: t('dashboard.rows-layout.row.new', 'New row') })];
+      rows = [new RowItem({ layout: layout.clone() })];
     }
 
     return new RowsLayoutManager({ rows });
   }
-
-  public static Component = ({ model }: SceneComponentProps<RowsLayoutManager>) => {
-    const { rows } = model.useState();
-    const styles = useStyles2(getStyles);
-
-    return (
-      <div className={styles.wrapper}>
-        {rows.map((row) => (
-          <row.Component model={row} key={row.state.key!} />
-        ))}
-      </div>
-    );
-  };
-}
-
-function getStyles(theme: GrafanaTheme2) {
-  return {
-    wrapper: css({
-      display: 'flex',
-      flexDirection: 'column',
-      gap: theme.spacing(1),
-      flexGrow: 1,
-      width: '100%',
-    }),
-  };
 }
