+++ public/app/features/dashboard-scene/scene/types/BulkActionElement.ts
@@ -1,5 +1,6 @@
 export interface BulkActionElement {
   onDelete(): void;
+
   onCopy?(): void;
 }
