+++ public/app/features/dashboard-scene/scene/types/LayoutParent.ts
@@ -6,6 +6,15 @@ import { DashboardLayoutManager } from './DashboardLayoutManager';
  * This interface is needed to support layouts existing on different levels of the scene (DashboardScene and inside the TabsLayoutManager)
  */
 export interface LayoutParent extends SceneObject {
+  /**
+   * Returns the inner layout manager
+   */
+  getLayout(): DashboardLayoutManager;
+
+  /**
+   * Switches the inner layout manager
+   * @param newLayout
+   */
   switchLayout(newLayout: DashboardLayoutManager): void;
 }
